requirements.txt
batchgenerators==0.25
fastcore==1.6.3
fran==0.9.1
ipdb==0.13.13
litq==0.1.0
matplotlib==3.10.1
networkx==3.3
numpy==2.2.3
pandas==2.2.3
pytest==8.3.3
pyxnat==1.6.2
ray==2.36.1
scipy==1.15.2
SimpleITK==2.4.1
SimpleITK==2.4.1
torch==2.5.1+cu121
utilz.egg==info
vtk==9.3.1
